Cleaning up some stuff
