+++ frontend/src/components/Heading/HeadingEditor/index.css
