+++ frontend/src/components/TextAnswer/TextInput.jsx
@@ -10,27 +10,8 @@ import {
   LOAD_QUESTION,
   SAVE_QUESTION,
 } from "../../redux/actions/questionActions";
-import {
-  SET_LOGIC_QUESTION,
-  SET_LOGIC_VIEW_QUESTION,
-} from "../../redux/actions/logicActions";
-import { TOGGLE_LOGIC } from "../../redux/actions/uiActions";
 
 function CustomerTextAnswer({ question }) {
-  const dispatch = useDispatch();
-  const logicList = useSelector((state) => state.logicReducer.logicList);
-  const questionList = useSelector(
-    (state) => state.questionReducer.questionList
-  );
-  const [questionNum, setQuestionNum] = useState("");
-  const [title, setTitle] = useState("");
-
-  useEffect(() => {
-    console.log(question);
-    setQuestionNum(`Q${question.position_index}`);
-    setTitle(question.question);
-  }, [question]);
-
   return (
     <div className="GlobalEditorComponent">
       <div className="GlobalEditorComponentHeader">
