+++ frontend/src/screens/request-form/request-form.js
@@ -3,11 +3,11 @@ import { useDispatch, useSelector } from "react-redux";
 import "./request-form.css";
 import { appColor } from "../../constants";
 import { LOAD_QUESTION } from "../../redux/actions/questionActions";
-import MultiSelect from "../../components/MultiSelect";
+import MultiSelect from "../../components/MultiSelect/MultiSelectEditor";
 import SingleSelect from "../../components/SingleSelect";
-import TextAnswer from "../../components/TextAnswer";
+import TextAnswer from "../../components/TextAnswer/TextAnswerEditor";
 import Dropdown from "../../components/Dropdown";
-import Heading from "../../components/Heading";
+import Heading from "../../components/Heading/HeadingEditor";
 import TextLine from "../../components/TextLine";
 import FileInput from "../../components/FileInput";
 import FileDownload from "../../components/FileDownload";
